pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "musical-perception"
version = "0.1.0"
description = "Extract structured musical parameters from audio — tempo, subdivision, exercise type, counting signature"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
]

[project.optional-dependencies]
whisper = [
    "openai-whisper>=20231117",
]
prosody = [
    "librosa>=0.10",
    "praat-parselmouth>=0.4",
    "soundfile>=0.12",
]
all = [
    "musical-perception[whisper,prosody]",
]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
